+++ backend/src/main/java/com/beyond/specguard/auth/controller/AuthController.java
@@ -1,6 +1,6 @@
 package com.beyond.specguard.auth.controller;
 
-import com.beyond.specguard.auth.model.dto.request.ClientSignupRequestDto;
+import com.beyond.specguard.auth.model.dto.request.SignupRequestDto;
 import com.beyond.specguard.auth.model.dto.request.InviteSignupRequestDto;
 import com.beyond.specguard.auth.model.dto.response.InviteCheckResponseDto;
 import com.beyond.specguard.auth.model.dto.response.ReissueResponseDto;
@@ -8,8 +8,8 @@ import com.beyond.specguard.auth.model.dto.response.SignupResponseDto;
 import com.beyond.specguard.auth.model.dto.response.TokenResponseDto;
 import com.beyond.specguard.auth.model.service.LogoutService;
 import com.beyond.specguard.auth.model.service.ReissueService;
-import com.beyond.specguard.auth.model.service.local.InviteSignupService;
-import com.beyond.specguard.auth.model.service.local.SignupService;
+import com.beyond.specguard.auth.model.service.InviteSignupService;
+import com.beyond.specguard.auth.model.service.SignupService;
 import com.beyond.specguard.auth.model.service.oauth2.TokenService;
 import com.beyond.specguard.common.util.CookieUtil;
 import com.beyond.specguard.common.util.TokenResponseWriter;
@@ -42,7 +42,7 @@ public class AuthController {
     private final TokenService tokenService;
 
     @PostMapping("/signup/company")
-    public ResponseEntity<SignupResponseDto> signup(@Valid @RequestBody ClientSignupRequestDto request) {
+    public ResponseEntity<SignupResponseDto> signup(@Valid @RequestBody SignupRequestDto request) {
         SignupResponseDto response = signupService.signup(request);
         return ResponseEntity.ok(response);
     }
