+++ backend/src/main/java/com/beyond/specguard/auth/model/dto/request/LoginRequestDto.java
@@ -6,12 +6,12 @@ import lombok.NoArgsConstructor;
 
 @Getter
 @NoArgsConstructor
-public class ClientLoginRequestDto {
+public class LoginRequestDto {
     private String email;
     private String password;
 
     @Builder
-    public ClientLoginRequestDto(String email, String password) {
+    public LoginRequestDto(String email, String password) {
         this.email = email;
         this.password = password;
     }
