+++ backend/src/main/java/com/beyond/specguard/auth/model/filter/ClientLoginFilter.java
@@ -1,8 +1,8 @@
 package com.beyond.specguard.auth.model.filter;
 
-import com.beyond.specguard.auth.model.dto.request.ClientLoginRequestDto;
-import com.beyond.specguard.auth.model.handler.local.CustomFailureHandler;
-import com.beyond.specguard.auth.model.handler.local.CustomSuccessHandler;
+import com.beyond.specguard.auth.model.dto.request.LoginRequestDto;
+import com.beyond.specguard.auth.model.handler.CustomFailureHandler;
+import com.beyond.specguard.auth.model.handler.CustomSuccessHandler;
 import com.beyond.specguard.auth.model.token.ClientAuthenticationToken;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import jakarta.servlet.http.HttpServletRequest;
@@ -34,7 +34,7 @@ public class ClientLoginFilter extends UsernamePasswordAuthenticationFilter {
                                                 HttpServletResponse response) throws AuthenticationException {
         try {
             ObjectMapper objectMapper = new ObjectMapper();
-            ClientLoginRequestDto loginDTO = objectMapper.readValue(request.getInputStream(), ClientLoginRequestDto.class);
+            LoginRequestDto loginDTO = objectMapper.readValue(request.getInputStream(), LoginRequestDto.class);
 
             UsernamePasswordAuthenticationToken authToken =
                     new ClientAuthenticationToken(loginDTO.getEmail(), loginDTO.getPassword());
