+++ backend/src/main/java/com/beyond/specguard/auth/model/service/InviteSignupService.java
@@ -1,4 +1,4 @@
-package com.beyond.specguard.auth.model.service.local;
+package com.beyond.specguard.auth.model.service;
 
 import com.beyond.specguard.auth.model.dto.request.InviteSignupRequestDto;
 import com.beyond.specguard.auth.model.dto.response.InviteCheckResponseDto;
