+++ backend/src/main/java/com/beyond/specguard/common/config/SecurityConfig.java
@@ -4,8 +4,8 @@ import com.beyond.specguard.auth.model.configurer.CommonSecurityConfigurer;
 import com.beyond.specguard.auth.model.filter.AdminLoginFilter;
 import com.beyond.specguard.auth.model.filter.ApplicantLoginFilter;
 import com.beyond.specguard.auth.model.filter.ClientLoginFilter;
-import com.beyond.specguard.auth.model.handler.local.CustomFailureHandler;
-import com.beyond.specguard.auth.model.handler.local.CustomSuccessHandler;
+import com.beyond.specguard.auth.model.handler.CustomFailureHandler;
+import com.beyond.specguard.auth.model.handler.CustomSuccessHandler;
 import com.beyond.specguard.auth.model.handler.oauth2.OAuth2FailureHandler;
 import com.beyond.specguard.auth.model.handler.oauth2.OAuth2SuccessHandler;
 import com.beyond.specguard.auth.model.provider.AdminAuthenticationProvider;
