+++ backend/src/main/java/com/beyond/specguard/invite/controller/InviteController.java
@@ -1,6 +1,6 @@
 package com.beyond.specguard.invite.controller;
 
-import com.beyond.specguard.auth.model.service.local.CustomUserDetails;
+import com.beyond.specguard.auth.model.service.CustomUserDetails;
 import com.beyond.specguard.invite.model.dto.request.InviteRequestDto;
 import com.beyond.specguard.invite.model.dto.response.InviteResponseDto;
 import com.beyond.specguard.invite.model.service.InviteService;
